// src/api/model/mailMessageDto.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * CRM.Backend.Web | v1
 * OpenAPI spec version: 1.0.0
 */

export interface MailMessageDto {
	/** @nullable */
	id?: string | null;
	/** @nullable */
	subject?: string | null;
	/** @nullable */
	from?: string | null;
	date?: string;
	/** @nullable */
	body?: string | null;
}